requirements.txt
flask==2.3.2
cs50==9.2.5
Flask-Session==0.8.0
Werkzeug==3.0.1
secure-smtplib==0.1.1
requests==2.31.0
Pillow==10.0.1
cloudinary==1.41.0
python-dotenv==1.0.1
Flask-APScheduler==1.12.4